+++ lib/http_wrapper.rb
@@ -1,6 +1,8 @@
+# frozen_string_literal: true
+
 require 'http_wrapper/version'
 require 'http_wrapper/constants'
 require 'http_wrapper/errors'
-require 'http_wrapper/utils'
+require 'http_wrapper/util'
 require 'http_wrapper/request'
 require 'http_wrapper/http_wrapper'