+++ lib/http_wrapper/errors.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class HTTPWrapper
   HTTPWrapperError      = Class.new StandardError
   TooManyRedirectsError = Class.new HTTPWrapperError