+++ lib/http_wrapper/request.rb
@@ -1,11 +1,13 @@
+# frozen_string_literal: true
+
 require 'uri/common'
 
 class HTTPWrapper
   class Request
-    KNOWN_PARAMS_KEYS = [:headers, :query, :cookie, :auth, :body, :user_agent, :content_type, :multipart].freeze
+    KNOWN_PARAMS_KEYS = %i[headers query cookie auth body user_agent content_type multipart].freeze
 
-    def initialize(url, method, params = {})
-      Utils.validate_hash_keys params, KNOWN_PARAMS_KEYS
+    def initialize(url, method, params = {}) # rubocop:disable Metrics/AbcSize
+      Util.validate_hash_keys params, KNOWN_PARAMS_KEYS
 
       self.uri = url
 
@@ -30,7 +32,7 @@ class HTTPWrapper
     attr_reader :uri
 
     def uri=(url)
-      url = "http://#{url}" unless url =~ /\Ahttps?:\/\//
+      url = "http://#{url}" unless url =~ %r{\Ahttps?://}
       @uri = URI.parse url
     end
 
@@ -43,16 +45,14 @@ class HTTPWrapper
 
     def normalize_headers(headers)
       normal_headers = {}
-      if headers
-        headers.each do |header, value|
-          normal_headers[normalize_header header] = value
-        end
+      headers&.each do |header, value|
+        normal_headers[normalize_header header] = value
       end
       normal_headers
     end
 
     def normalize_header(header)
-      header = header.to_s.gsub(/_/, '-') if header.is_a? Symbol
+      header = header.to_s.tr('_', '-') if header.is_a? Symbol
       header.downcase
     end
 
@@ -62,28 +62,27 @@ class HTTPWrapper
 
     def default_content_type_for(method)
       case method
-        when :post, :put then CONTENT_TYPE::POST
-        else CONTENT_TYPE::DEFAULT
+      when :post, :put then POST_CONTENT_TYPE
+      else DEFAULT_CONTENT_TYPE
       end
     end
 
     def initialize_headers
-      @headers[HEADER::USER_AGENT]   ||= @user_agent
-      @headers[HEADER::CONTENT_TYPE] ||= @content_type
-      @headers[HEADER::COOKIE]       ||= @cookie if @cookie
+      @headers[USER_AGENT_HEADER_NAME]   ||= @user_agent
+      @headers[CONTENT_TYPE_HEADER_NAME] ||= @content_type
+      @headers[COOKIE_HEADER_NAME]       ||= @cookie if @cookie
     end
 
     def merge_uri_query
-      return unless @query.size > 0
-      original_query = @uri.query ? Utils.query_to_hash(@uri.query) : {}
+      return if @query.empty?
+
+      original_query = @uri.query ? Util.query_to_hash(@uri.query) : {}
       merged_query = original_query.merge @query
-      @uri.query = Utils.hash_to_query merged_query
+      @uri.query = Util.hash_to_query merged_query
     end
 
     def create_method_specific_request
-      # Ruby v1.9.3 doesn't understand full URI object, it needs just path :(
-      uri = RUBY_VERSION =~ /\A2/ ? @uri : @uri.request_uri
-      @request = @method.new uri, @headers
+      @request = @method.new @uri, @headers
       set_body
       set_basic_auth
       @request
@@ -91,6 +90,7 @@ class HTTPWrapper
 
     def set_body
       return unless @request.request_body_permitted?
+
       if @multipart_data
         set_body_from_multipart_data
       else
@@ -100,21 +100,23 @@ class HTTPWrapper
 
     def set_body_from_multipart_data
       convert_body_data_to_multipart_data if @body_data
-      @request.set_form @multipart_data, CONTENT_TYPE::MULTIPART
+      @request.set_form @multipart_data, MULTIPART_CONTENT_TYPE
     end
 
     def convert_body_data_to_multipart_data
-      @body_data = Utils.query_to_hash(@body_data) unless @body_data.kind_of? Hash
-      @body_data.each{|key, value| @multipart_data << [key.to_s, value.to_s]}
+      @body_data = Util.query_to_hash(@body_data) unless @body_data.is_a? Hash
+      @body_data.each { |key, value| @multipart_data << [key.to_s, value.to_s] }
     end
 
     def set_body_from_body_data
       return unless @body_data
-      @request.body = @body_data.is_a?(Hash) ? Utils.hash_to_query(@body_data) : @body_data
+
+      @request.body = @body_data.is_a?(Hash) ? Util.hash_to_query(@body_data) : @body_data
     end
 
     def set_basic_auth
-      return unless @login
+      return unless @login && @password
+
       @request.basic_auth @login, @password
     end
   end